# disable Terser's "reduce_funcs" option for performance