<!--- Provide a general summary of your changes in the Title above -->
Terser's  “reduce_funcs” option seems to degrade performance. So, disable it. This commit will increase file size of "mithril.min.js" slightly, but performance appears to improve.

## Description
<!--- Describe your changes in detail -->
~~This PR makes mithril.js and mithril.min.js code equivalent (except mangling) by disabling Terser's “compress” option.~~
This will increase the file size of “mithirl.min.js” by about 0.01KB (gzipped), but seems to improve performance.
**Edit:** The code in mithril.js and mithril.min.js is not equivalent because only reduce_funcs is disabled.  

results of `npm run perf` (nodejs, using mithril.min.js)
- before
```
construct large vnode tree x 21,998 ops/sec ±0.49% (127 runs sampled)
rerender identical vnode x 3,876,899 ops/sec ±0.47% (122 runs sampled)
rerender same tree x 83,591 ops/sec ±0.19% (126 runs sampled)
add large nested tree x 6,869 ops/sec ±1.55% (122 runs sampled)
mutate styles/properties x 123 ops/sec ±1.29% (91 runs sampled)
repeated add/removal x 3,007 ops/sec ±0.69% (124 runs sampled)
Completed perf tests in 56568ms
```

- after (improves "rerender same tree" and "add large nested tree")
```
construct large vnode tree x 20,641 ops/sec ±0.51% (128 runs sampled)
rerender identical vnode x 4,029,056 ops/sec ±0.56% (129 runs sampled)
rerender same tree x 90,134 ops/sec ±0.25% (125 runs sampled)
add large nested tree x 7,452 ops/sec ±0.84% (123 runs sampled)
mutate styles/properties x 130 ops/sec ±1.41% (93 runs sampled)
repeated add/removal x 3,047 ops/sec ±0.34% (128 runs sampled)
Completed perf tests in 57709ms
```

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
While measuring performance under various conditions, I noticed the performance was different between "mithril.min.js" and "mithril.js" ("mithril.min.js" seems slower).
Looking into the cause, I found that compressing the code with Terser's “compress” option seemed to affect performance as well as it reduced the amount of code.

I use mithril.min.js as is for prototyping, so I prefer mithril.min.js to be faster.

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
`npm run test`

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [ ] My change requires a documentation update, and I've opened a pull request to update it already:
- [x] I have read https://mithril.js.org/contributing.html.