@dead-claudia 2024-12-25
I'm 90% sure [it's related to the fact Terser inlines larger functions by creating IIFEs it doesn't proceed to flatten](https://github.com/terser/terser/issues?q=is%3Aissue+is%3Aopen+iife).

Better would be finding the offending function and either inlining it manually (if it's easy) or marking it with `/*@__NOINLINE_*/` (if it's complicated).