@kfule 2024-12-26
@dead-claudia Thank you for your review comments.
In my opinion, I don't want to make changes to the mithril code itself because of terser's behavior. 
So I want to improve only terser options. But it seems that the version of terser needs to be upgraded for reduce_funcs to work well, and it may not support node 18...