@kfule 2024-12-26
@dead-claudia 
Very sorry, the node 18 issue was not related to the Terser's upgrade.
So, I upgraded Terser to make reduce_funcs work well. This allowed better performance with little increase in file size (0.01KB gzipped).